pyproject.toml
[project]
name = "cross-check"
version = "0.1.0"
description = "A multi-agent system for cross-checking spam URLs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "google-adk[eval]>=1.19.0",
    "litellm>=1.80.7",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "requests>=2.32.5",
    "validators>=0.35.0",
    "markitdown>=0.0.1",
    "mesop>=1.1.1",
]
